chore: seperate api app from scraper
