+++ apps/scraper-api/.env.template
@@ -1,0 +1,1 @@
+API_SECRET_KEY=
