+++ apps/scraper-api/.eslintrc.js
@@ -1,0 +1,7 @@
+module.exports = {
+  root: true,
+  extends: ['../../.eslintrc.base.js', 'plugin:import/recommended'],
+  rules: {
+    'no-console': 'off',
+  },
+}
