+++ apps/scraper-api/lint-staged.config.js
@@ -1,0 +1,2 @@
+const config = require('../../lint-staged.config.base')
+module.exports = config
