+++ apps/scraper-api/package.json
@@ -1,30 +1,21 @@
 {
-  "name": "conbini-this-week-scraper",
+  "name": "conbini-this-week-scraper-api",
   "version": "0.0.0",
   "private": true,
   "scripts": {
     "fix:staged-files": "lint-staged --allow-empty",
     "lint": "eslint . --ext ts",
-    "scrape": "ts-node -r tsconfig-paths/register src/index.ts",
     "type-check": "tsc --pretty --noEmit"
   },
-  "dependencies": {
-    "jsdom": "^21.0.0",
-    "zod": "^3.20.2"
-  },
   "devDependencies": {
-    "@types/jsdom": "20.0.1",
     "@types/node": "18.11.18",
     "@typescript-eslint/eslint-plugin": "5.48.1",
     "@typescript-eslint/parser": "5.48.1",
     "@vercel/node": "2.8.15",
-    "dotenv": "16.0.3",
     "eslint": "8.31.0",
     "eslint-config-prettier": "8.6.0",
     "eslint-import-resolver-typescript": "3.5.3",
     "eslint-plugin-import": "2.27.4",
-    "ts-node": "10.9.1",
-    "tsconfig-paths": "4.1.2",
     "typescript": "4.9.4"
   }
 }
