+++ packages/scraper/src/scrape.ts
@@ -2,7 +2,7 @@ import { z } from 'zod'
 import { Client } from '../../db/src'
 import type { InsertItem } from '../../db/src/types'
 import { supabaseKey, supabaseUrl, conbinisConfig } from './constant'
-import { ConbiniName, conbiniNames } from '@conbini-this-week/core'
+import { ConbiniName, conbiniNames } from '../../core/src'
 import { JSDOM } from 'jsdom'
 
 export async function scrape(name: ConbiniName) {
